+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - missing parameters to Boleto, BrcodePayment, Deposit, DictKey and Invoice resources
 - InvoicePayment sub-resource to allow retrieval of invoice payment information
 - Event.Attempt sub-resource to allow retrieval of information on failed webhook event delivery attempts
+- pdf method for retrieving PDF receipts from reversed invoice logs
 
 ### Fixed
 - special characters in BrcodePreview query